refactor(trusty-search): split 3 files under SLOC caps (#1195)
Behavior-preserving refactor to bring three over-cap trusty-search files
below their SLOC limits (production 500, test 1500).

- core/indexer/tests.rs (2472 SLOC) -> tests/ directory: shared fixtures in
  mod.rs, tests grouped by concern into persistence_and_search,
  ranking_and_modes, branch_and_corpus, eviction_kg_paths (each <700 SLOC).
  All 74 tests preserved; make_indexer_with_corpus promoted to mod.rs as it
  is shared by two concern modules.
- service/call_chain.rs (738 SLOC) -> call_chain/ dir: production in mod.rs
  (430 SLOC), inline tests moved to call_chain/tests.rs (300 SLOC).
- service/embedder_supervisor.rs (521 SLOC) -> embedder_supervisor/ dir:
  production in mod.rs (295 SLOC), inline tests to tests.rs (225 SLOC).

Public paths unchanged (re-exported via mod.rs); no external refs move.
Drops the three resolved entries from .line-cap-allowlist.tsv.
