+++ .line-cap-allowlist.tsv
@@ -15,6 +15,3 @@ crates/trusty-analyze/src/main.rs	567
 crates/trusty-analyze/src/mcp/mod.rs	892
 crates/trusty-gworkspace/src/tools.rs	557
 crates/trusty-memory/src/lib.rs	618
-crates/trusty-search/src/core/indexer/tests.rs	2472
-crates/trusty-search/src/service/call_chain.rs	738
-crates/trusty-search/src/service/embedder_supervisor.rs	521
